canio: rename from _canio
This reflects our belief that the API is stable enough to avoid incompatible changes during 6.x.
