+++ shared-module/canio/Match.c
@@ -24,14 +24,20 @@
  * THE SOFTWARE.
  */
 
-#pragma once
+#include "shared-module/canio/Match.h"
 
-#include "py/obj.h"
-#include "shared-module/_canio/Match.h"
+void common_hal_canio_match_construct(canio_match_obj_t *self, int address, int mask, bool extended) {
+    self->address = address;
+    self->mask = mask;
+    self->extended = extended;
+}
 
-extern const mp_obj_type_t canio_match_type;
-
-void common_hal_canio_match_construct(canio_match_obj_t *self, int address, int mask, bool extended);
-int common_hal_canio_match_get_address(const canio_match_obj_t *self);
-int common_hal_canio_match_get_mask(const canio_match_obj_t *self);
-bool common_hal_canio_match_get_extended(const canio_match_obj_t *self);
+int common_hal_canio_match_get_address(const canio_match_obj_t *self) {
+    return self->address;
+}
+int common_hal_canio_match_get_mask(const canio_match_obj_t *self) {
+    return self->mask;
+}
+bool common_hal_canio_match_get_extended(const canio_match_obj_t *self) {
+    return self->extended;
+}
