+++ shared-module/canio/Message.h
